ish: max-retries mechanism for watchdog timer
This adds relevant config options, as well as implementation for a
max-retries mechanism on the watchdog timer. Included is an
implementation for ISH, but should be easy to adapt to other boards
with the save/restore functions that can be implemented on a
chip-by-chip basis.

BUG=b:132059981
BRANCH=none
TEST=observed watchdog timer disable after 4 resets, then re-enable
once we had a successful reset

Change-Id: I7b443d9a20a474b294d494c5b6046a38eaf6ff12
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>