Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> ToT has really diverged from the branch wrt to mkbp_event. These were all the CLs I needed to cherry-pick to get it to parity.

Understood. I assumed that we would not necessarily care to keep branches up-to-date with master unless we had a reason. These CLs are mostly refactoring in support for HECI, not anything that would specifically improve the performance on a GPIO system.

If there is some overarching goal why we need this parity, then by all means, I am happy to +2, but I'd like to understand the rationale to introduce change on a shipping and working system.

Patch-set: 1
Reviewer: Gerrit User 1287098 <1287098@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>