kohaku: Enable CONFIG_CHARGER_BQ25710_IDCHG_LIMIT_MA
This config option is used to both set a limit for IDCHG and enable
IDCHG as a trigger for prochot. With this feature enabled, Kohaku is
able to boot into the kernel and remain running even under heavy loads
while on battery only. The initial battery that's being used is only
2S+1P which does not provide sufficient power for CPU peak load
conditions.

BUG=b:132285560
BRANCH=none
TEST=Manual
Verfied that Kohaku can boot running on battery only and remains
powered up as the CPU load gets increased.

Change-Id: Idfc4b65d2b86ee9883345c6aa610130d2b9fe89e
Signed-off-by: Scott Collyer <scollyer@google.com>