flash_ec: expand the list of directories where npcx_monitor.bin can be found
Add the dirs for LOCAL_BUILD or EMERGE_BUILD to the list of directories
to search for npcx_monitor.bin. Prioritize the directory where the EC
binary is located (which could be specified with the --image parameter).

BUG=b:132350402
BRANCH=none
TEST=Build with 'sudo emerge ec-devutils'
Test with and without an --image parameter:
 flash_ec --board=kohaku
 flash_ec --board=kohaku --image=build/kohaku/ec.bin
Verify that flash_ec is able to locate npcx_monitor.bin in both cases.

Change-Id: I5a679d1bc215cc0383f60e02aa702db79cd968b3
Signed-off-by: Paul Fagerburg <pfagerburg@chromium.org>