Update patch set 2
Patch Set 2:

Talking to Jett, I'm going to re-work this just a bit to also have the suspend state respond to the sysjump event.  

There's some small chance we could get a PD interrupt in between checking the PD task state and setting the event, and then if the PD RX read fails the port will be suspended before this task runs again and sets the event.  It's a very small chance and definitely on the paranoid side, but seems worthwhile given the consequence of not fixing it is hanging the host command task. Alternatively, we could set some sufficiently long timeout to wait on the return event, though if someone else writes a tight wait loop into the PD task the timeout would mask that the loop may need to handle sysjump-related prep.

Patch-set: 2
Reviewer: Gerrit User 1287197 <1287197@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0