Update patch set 3
Patch Set 3:

> i2c_bus & port: Retrieved by EC_CMD_I2C_LOOKUP in one call. EC_CMD_I2C_LOOKUP takes a chip class as a parameter. We can make it return all (i2c_bus, port) pairs which belong to the TCPC class. Thus it covers the port count (EC_CMD_USB_PD_PORTS). This update also benefits other users of EC_CMD_I2C_LOOKUP.

So how do you get the PD port number that PD_CHIP_INFO takes as an input, then? It doesn't seem like they are always numbered in order (e.g. octopus seems to have a PD port 1 but not a PD port 0).

I'm okay with combining this into I2C_LOOKUP if you want, but then that command needs to return enough info so the result is actionable on its own. Why not just expand it to include a few bytes of class-specific information and put the PD port number in there for TCPCs? (And since it was only just added and no EC shipped with it yet, you don't even need to version it for that change.)

Patch-set: 3
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>