common: move reset flag definitions into include file
This came up in a code review that we need an easy way to check and
see if a given set of reset flags was valid. Without hard coding the
number of reset flags in a chip-specific file, there was no easy way
to do this.

This commit refactors the reset flag definitions to use a new
reset_flags.inc file, which gets enumerated and turned into flags in
system.h, and the descriptions stored into an array in
common/system.c (as well as ectool).

This helps make our code safer and more extensible, as now we can do a
check to make sure we have no more flags than we can store, as well as
getting definitions like RESET_FLAGS_COUNT. In addition, it's easier
to define new reset flags as needed.

BUG=none
BRANCH=none
TEST=make buildall -j
TEST=made sure reset flags acted as normal on arcada_ish

Change-Id: Ic72622b16af5aeeb13c52de17170122e028d36af
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>