gsctool: refactor command line arguments processing
There are a few issues with command line processing code:

  - command line options must be listed in two different places if we
    want to allow both short and long forms

  - help text for commands is included in yet another place in the code

  - short command description string does not properly identify
    commands with optional arguments (where the short command option
    is supposed to be followed by ::)

Indeed, from the libc getopt_long() point of view the long option
table and short options string are completely unrelated: the function
scans the command line arguments vector for -x[x[y]] or --word command
line options style and looks up either short_opts or long_opts
depending on the encountered command line token's style.

The standard way to cross reference long options table and the short
options string is to put the short option character in the long option
table's entries .val field. This is how it is done in gsctool.

What this means is that as long as all short command options are
referenced in the long_opts table, the short command option string
could be generated based no the table.

The help text and long_opts table still remain separate and require
manual syncing.

This patch introduces a container structure, which includes both long
option contents and the help text for each command line option.

This allows to build the long_opts table and then the short_opts
string on the fly, based on the contents of the single table
describing command line options including their help text.

BRANCH=none
BUG=none
TEST=as follows:

   - temporarily added debug code to print out the generated short
     option string, got the following:

  aBbcd:fF:hIi::kL::Mm::n:O:oPpR:r::S:stUuVvw

    which is extended to show optional arguments compared to the
    short_opts string being removed by this patch:

  aBbcd:F:fhIikLMmn:O:oPpR:rS:stUuVvw

  - verified that 'gsctool -h' outputs before and after this patch are
    exactly the same

Change-Id: Ie8d409c6c8866247323cee93ce5b9bfbe22046fa
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>