gsctool: refactor command line arguments processing
From the libc getopt_long() point of view the long option table and
short options string are completely unrelated: the function scans the
command line arguments vector for -x[x[y]] or --word command line
options style and looks up either short_opts or long_opts depending on
the style.

The standard way to connect long options table and the short options
string is to put the short option character in the long option table's
entries .val field. This is how it is done in gsctool.

This creates a maintenance issue, each time a new command line option
is added it needs to be included in both long and short forms
separately.

Since we don't have short options not present in one of the .val
fields of the long option table, we can build the short option string
at run time based on the long option table contents, before
getopt_long() is invoked.

The long_opts table contains information about optional command line
parameters of certain command line arguments, saved in the .var field
of the table entries.

But getopt_long() does not set the index field value for when the
command line option is used in the short form.

This patch adds a function which maps the sort command line option
retrieved by getopt_long() into the long_opts table entry, so that the
.has_arg field could be examined to check if an optional command line
argument is possible.

This allows to retrieve the pointer to the command line argument in
one place instead of repeating the same code when processing all
relevant options.

BRANCH=none
BUG=none
TEST=added debug code to print out the generated short option string,
     got the following:

  aBbiIkoUcd:LF:fhMO:PprS:R:sn:mtVvwu

    which is logically identical to the short_opts string being
    removed by this patch:

  aBbcd:F:fhIikLMmn:O:oPpR:rS:stUuVvw

Change-Id: Ie8d409c6c8866247323cee93ce5b9bfbe22046fa
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>