Update patch set 3
Patch Set 3:

I think there are still some problematic details with this approach. Besides updating, our current TCPC software sync code also calls EC_CMD_I2C_PASSTHRU_PROTECT on every TCPC I2C bus individually even when it's not updating... and that's a problem because that needs the I2C bus number (not the PD port number) which we now no longer have in the normal boot (no update) case. Do we still need that behavior? It looks like we want to protect all the busses at the end anyway (what CL:1605260 is trying to do). So can we just take the individual protection code out from all the AUX FW drivers? (That means getting rid of the .protect() function pointer in VbootAuxFwOps and all implementations of it.)

The other problem is that we're checking whether the bus is protected in check_vboot_aux_fw() if we decide to update the TCPC so we can reboot the EC earlier if necessary (see CL:1605260), and that also needs the I2C bus number. Not sure if that optimization is really doing anything for any real scenario? In general, I think we either have:

1. EC was already in RO (e.g. TCPC update together with EC FW update). In that case, bus is guaranteed unprotected and we won't need another reboot.
2. EC is in RW but we still have a TCPC update. In that case, hardly anything happens between check_vboot_aux_fw() and update_vboot_aux_fw() because the EC is already in RW (e.g. we're not actually software-syncing). So what does it matter whether we detect that we need to reboot to unlock the I2C bus in the check or in the update phase? It doesn't seem like a meaningful time save compared to the TCPC update time.

Can we take that one out too? That would mean we can take out the protect_status() function pointer from VbootAuxFwOps (and all implementations) as well. Then we only have check_hash() and update_image() left, and check_hash() can be served with just the result from the EC_CMD_PD_CHIP_INFO that we already have. Then it should all work out.

Patch-set: 3
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>