Update patch set 3
Patch Set 3:

> So can we just take the individual protection code out from all the AUX FW drivers? (That means getting rid of the .protect() function pointer in VbootAuxFwOps and all implementations of it.)

Yes, with the CL to protect all the TCPC tunnels, we can remove the protect operation from the VbootAuxFw Driver.



> 2. EC is in RW but we still have a TCPC update. In that case, hardly anything happens between check_vboot_aux_fw() and update_vboot_aux_fw() because the EC is already in RW (e.g. we're not actually software-syncing). So what does it matter whether we detect that we need to reboot to unlock the I2C bus in the check or in the update phase? It doesn't seem like a meaningful time save compared to the TCPC update time.
> 
> Can we take that one out too? That would mean we can take out the protect_status() function pointer from VbootAuxFwOps (and all implementations) as well. Then we only have check_hash() and update_image() left, and check_hash() can be served with just the result from the EC_CMD_PD_CHIP_INFO that we already have. Then it should all work out.

If the EC does not have any updates & is already in RW, then the tunnels would be protected at the end of the previous RW jump. So when TCPC FW needs to be updated, we need protect_status operation to check the status of the tunnel and ask EC to reboot to RO. Please feel free to correct me if I misunderstood the existing design or your statement.

Patch-set: 3
Reviewer: Gerrit User 1302633 <1302633@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>