Update patch set 12
Patch Set 12:

> Patch Set 12: Code-Review-1
> 
> Any hope in splitting this CL in 2? It's really indicative of that when your CL's description starts with "This change does n things".
> 1. The spreading on the EC in one CL, ideally in common code.
> 2. interrupt GPIO retriggering, in the lsm6 driver.

Agreed that the spreading is common code. If that is the only thing blocking this CL, I am happy to volunteer to split the spreading in a follow-up CL.

Would it make sense to merge this in the Wilco FW branch and gate merging back to master on this refactoring work?

Patch-set: 12
Reviewer: Gerrit User 1287098 <1287098@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>