Update patch set 12
Patch Set 12:

It's just that I found stuff is a lot easier to review if it's organized. It's not great to review something before a huge refactor (both for readability point of view and bugs that might show up in the refactoring). Not to mention that some individual CLs from after a split might be able to go in before the rest.

Patch-set: 12
Reviewer: Gerrit User 1113904 <1113904@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>