FIXUP: sensor: Adjust max_frequency based on EC performance
When chaning the macro fro BMA255, the step was set to 125Hz, so it
assumes the EC can support MEMS ODR set to 125Hz.
This is not the case on poppy and nautilus in the poppy branch.

BRANCH=poppy
BUG=b:118205424,b:118851581,chromium:615059,b:131705379
TEST=Compile

Change-Id: Ib60abb3919dc1ce049211d848fc4e4de2e5e51c0
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1621188
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Enrico Granata <egranata@chromium.org>