driver: lsm6dsm: Fix missing events
BUG=b:129159505, b:195700255
BRANCH=None
TEST=Ran Android CTS
Ran tast run <ip> hardware.SensorIioserviceHard in loop.

Fixes missing event errors in CTS. This is done by updating the
last_fifo_read_ts and checking the interrupt GPIO. If we find that
the GPIO is still low at the end of the read, that means that we've
gotten new data while reading and never fully empties the FIFO. We
know this must have happened some time between the time we read the
count and when we were reading the FIFO, for now we'll use the upper
bound of this to be safe.

Change-Id: I0461f9d2703a3801e57e7769fbfe0e8de750706a
Signed-off-by: Yuval Peress <peress@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1620791
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>
(cherry picked from commit 1a98cd7c32a881331828bc975be069b7a30d8eb4)
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>