Update patch set 1
Patch Set 1:

General comment. Is anything about this specific to the LSM part?

It seems like we could do spreading-on-the-EC in general for all sensors? If so, would it make sense to write a CL for "sensor event spreading" abstracted from the specific sensor, and then have the LSM opt into that in a separate, much smaller, CL?

Patch-set: 1
Reviewer: Gerrit User 1287098 <1287098@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>