Update patch set 3
Patch Set 3:

> If the EC does not have any updates & is already in RW, then the tunnels would be protected at the end of the previous RW jump. So when TCPC FW needs to be updated, we need protect_status operation to check the status of the tunnel and ask EC to reboot to RO. Please feel free to correct me if I misunderstood the existing design or your statement.

Yes, we need to check the protection status when we update, but this doesn't need to be an extra VbootAuxFwOps callback. We're already doing that check and handling a protected EC correctly in the .update_image() callback. That's fine, I just want to get rid of the separate .protect_status() callback that Jett introduced in CL:1220428 (because I think it would create problems with this new flow). It sounds like it's just an optimization and as far as I can tell it doesn't really optimize anything important... so unless anybody (Jett?) knows why we need that, I'd propose to just get rid of it.

Patch-set: 3
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>