bobba: add blorb and droid into the support of controlling USB Type-A power.
Blorb and droid are first projects we are tyring to enable the control
of USB Type-A power in the run time. This CL set
inhibit_charging_in_suspend to USB_DISALLOW_SUSPEND_CHARGE in default
and relies on OS to change this flag based on the status of USB device
on the USB Type-A port.

BUG=b:130767435
BRANCH=octopus
TEST=1. `make BOARD=bobba -j4`
     2. check power of USB Type-A port is disabled when OS gets into
        S0iX and no any USB device is ever connected to DUT.

Change-Id: I19ff69df89a857cd3feb594641b6562f20356879
Signed-off-by: Marco Chen <marcochen@chromium.org>