Update patch set 3
Patch Set 3:

> 2. Change the code you quoted to not disable the display when rebooting from ec_sync_update_aux_fw().

Wait, that doesn't work. Looks like we disable the display intentionally there because for some TCPCs (ANX3429) we always want to reboot the EC after successful update.

Okay, other plan: it's fine to lookup the I2C tunnel in the check phase *iff* we have already determined that we need an update anyway. So the current .protect_status() usage should be okay, we just need to make sure that we never call protect_status() unless we have determined that we need an update (which is true in the current code flow). The I2C tunnel should then be lazily initialized on-demand by the first function that needs it.

Patch-set: 3
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>