core/host/task: Add OS sleeps instead of busy-looping
When running on a busy builder, fuzzing tests often time out.
I suspect this is due to the fact that one of the thread is busy
looping, and for some reason, does not give enough time for
other threads to run.

Fix the issue by using nanosleep OS call (avoiding namespace
clash with msleep/usleep that are implemented differently).

From the builder stack traces, we know these tests often
fail in 2 busy-loops, so fix those.

BRANCH=none
BUG=chromium:963768
TEST=Emulate busy system with 4 cores only:
     taskset 0xf yes > /dev/null & (about ten times)
     Then run tests:
     while taskset 0xf build/host/usb_pd_fuzz/usb_pd_fuzz.exe \
            -runs=1; do :; done
     => No failures after 30 minutes.

Change-Id: I458ff783d166e27fb38dc33853f08e5b3acba980
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>