Adds filtering of the STM32 DFU devices to eliminate devices with DFU's 'Runtime' identifier.
Some USB devices include the DFU Runtime descriptor during normal use.
These devices are detected as potential targets by DFU-UTIL causing it to
incorrectly identify multiple potential DFU targets leading to errors
while flashing. This change adds an additional requirement using the
device's serial identifier. The serial changes depending on the platform
so the device list is used to identify the device in DFU mode and it's
serial identifier. This serial is passed as an identifier to select the
correct device.

BUG=b:133329195
BRANCH=none
TEST=Verified DFU with STM32 based devices: Servo_v4, Servo_Micro, and
Sweetberry flash as expected. Verified that the awk correctly extracts
and eliminates devices with the DFU's Runtime identifier when a
suitable device connects. Verified pattern matching on variations of
the serial.
Signed-off-by: Brian Nemec <bnemec@chromium.org>
Change-Id: I2331b5e06d2eebf8dc28f20b9dd6b10bf2abe02a