cr50: console command 'intap' to adjust INT_AP_L pulse width.
intap can set and display the amount of INT_AP_L pulse width.
The unit is a loop count, which can corresponds to micro seconds
roughly.

BUG=b:130515803
BRANCH=None
TEST=tested on bob.

> intap
0 loop account

> intap 10
10 loop account

With 0, 0.5 usec pulse width was observed in logic analyzer.
With 3, 2.8~3.9 usec pulse width was observed in logic analyzer.
With 10, 7~9 usec pulse width was observed in logic analyzer.

Change-Id: I4ea1c45a64590321fc7e80468d18863b0990f9dc
Signed-off-by: Namyoon Woo <namyoon@chromium.org>