Casta: Disconnect C0 USB mux during forced shutdowns
During forced shutdowns (ex. EC console "apshutdown" or servod
"dut-control power_state:off"), the board chipset_do_shutdown code
will run before the PD task's hooks to disconnect sinks and mux
lines.  As a result, the USB mux may still be set to USB or DP.

On casta, the PS8751 is used as the mux for port 0 and draws more
current than we can support while turning off the 3.3V rail.  For this
specific scenario, we need to set the USB mux to "none" or else the
board may reboot during the shutdown.

BUG=b:129035454
BRANCH=octopus
TEST=ran "dut-control power_state:off" with several dongles connected
and verified the PP3300_EC dip was at most to 2.94V (with 47k R820
re-work on a Bluebird EVT)

Change-Id: I0ac00e92eaba1f3ca8854c1b8c925bc1b01b6ceb
Signed-off-by: Diana Z <dzigterman@chromium.org>