ec: common: Replace flag -D to -D= to be supported by IS_ENABLED().
IS_ENABLED works for an empty-string-defined macro.
However, -D options default to define the macro to 1.  This CL forces
those macros, such as BOARD_* CHIP_*, CORE_*, CHIP_VARIANT_* and
CHIP_FAMILIY_*, to be defined as an empty string, so that it can
be supported by IS_ENABLED macro.

TEST=use if(IS_ENABLED(BOARD_KRANE)) and see compilation success.
TEST=compares build directory w/ and w/o this CL, and see the .smap
     are the same:
     ls build/*/*/ec.*.smap | sed -e 's|build/||' | \
     xargs -I{} diff -u -a build/{} build.new/{}
BUG=none
BRANCH=None

Change-Id: I96e2aa1cb5f3369e5e445a674595a9234f26707a
Signed-off-by: Yilun Lin <yllin@google.com>