driver: add polling mode to TCS3400 sensor
Krane does not reserve an interrupt pin for ALS, so we need to poll the
sensor. Add a new config option CONFIG_ALS_TCS3400_POLLING to enable
polling mode.

Delay time is set to zero because the RGBC_VALID status flag seems to be
always true after the first data arrived.

BUG=b:129419982
BRANCH=None
TEST=Verify that `ectool motionsense` outputs als data on Krane

Change-Id: I960df249d29c0ac21810057e25f14d4bac3e14f5
Signed-off-by: Ting Shen <phoenixshen@google.com>