Updates the Sweetberry I2C speeds so they match measurements and are within spec.
All 4 I2C ports on the Sweetberry were configured at 800 KHz. 3 of the I2C
ports are limited to 400 KHz and the FMP I2C port can be up to 1 MHz. This
means that 3 of the slower speed I2C ports are out of spec by a factor of
two. Traces of the rails showed that the rails were still functioning at
800 KHz, although the FMP I2C rail is actually at 896 KHz. The STM32F4
Reference Manual Rev4 does not indicate which potential problems occur
when operating in this condition.

The I2C speeds were adjusted to match spec with the 3 slow speed ports set
to 400 KHZ. The I2C prescalers for the FMP I2C port do not have a suitable
1 MHz configuration so the FMP I2C rail was adjusted to 900 KHz to match
the measured speed within 2% for all rails. The cost of this change is the
scan time for a group of 28 INA's increased from 4.9 ms to 5.5 ms.

BRANCH=Servo
BUG=b:133503970
TEST=Verified Sweetberry can measure rails when used with powerlog.
Observed I2C rail activity using a Saleae analyzer.
Signed-off-by: Brian Nemec <bnemec@chromium.org>

Change-Id: Id345c1dd902212b92089e5ab3e04dd187efb4542