Update patch set 2
Patch Set 2: Verified+1 Commit-Queue+2

> Patch Set 2: Code-Review+1
> 
> Out of curiosity, why would some boards differ in supplier priority? I'm not familiar with the WPC suppliers, so maybe it has something to do with that, but the remaining suppliers have a defined priority in the USB specs.

It depends on the device policy and the power topology. WPC and USB are different spec. So, we are the one who decides the priority. Ideally, we want to make WPC and USB sources compete: we pick whichever offers the highest power but the hardware has to support that. On flapjack, WPC is forced to shut down when USB power is present.

Patch-set: 2
Reviewer: Gerrit User 1106949 <1106949@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+2
Label: Verified=+1