tcpc: Consider WPC status to determine VBUS
PROOF_OF_CONCEPT

Since VBUS is shared by WPC and USB and TCPC lives in the charger,
TCPC triggers an alert on VBUS detection when WPC is connected.

EC has to judge whether it's from WPC or USB but it cannot be done
without knowledge about the power topology.

So, simply allowing boards to define their own priority (CL:1622889)
is a cleaner solution.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=none
BRANCH=none
TEST=none

Change-Id: I65ce1a942ca6382bc14876910ab2be52e63b664b