common: board_[read/write]_serial weak reference cleanup
board_read_serial and board_write_serial were prototyped as weak
and this made all instances, that included that prototype, weak
as well.  In order to not lose information from the prototype,
default and override functions, I changed to use the override
weak marker symbols.

These functions defaulted for specific configurations as
different functionality and used an #ifdef tree to do this. I
made these a single definition for each function and used
IS_ENABLED instead of the #ifdef tree.  I also added a
definition for the case that the configuration would not have
produced a function.

BUG=none
BRANCH=none
TEST=make buildall -j

Change-Id: Ie41c53f3a17d665358e46eefd3ded3066ee80a7d
Signed-off-by: Denis Brockus <dbrockus@chromium.org>