cr50: Remove user presence check message.
Coreboot on Wilco now uses phyiscal presense checks from the Cr50 to
confirm user actions.  When booting to the recovery screen, the Cr50
console is flooded with "Presence: 0" messages.  Demote the message to
DBG images only and only if the user presense was detected.

BUG=b:133266946
BRANCH=cr50
TEST=Build DBG and non DBG CR50 images on Wilco. Confirm the message
flood is gone at the recovery screen.

Change-Id: Ie6b08a386e9e479e61a6764dad8d4a3c9aa57f05
Signed-off-by: Keith Short <keithshort@chromium.org>