Update patch set 15
Patch Set 15:

If we're already redesigning the whole command anyway, why don't we change it so that it will return records for all chips of the requested type? That would be more efficient (just one command in critical path rather than number of TCPCs + 1) and probably more useful (not all chip types we might add here in the future will necessarily have an out-of-band mechanism to check how many instances exist).

Patch-set: 15
CC: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>