Update patch set 15
Patch Set 15:

> Patch Set 15:
> 
> > Patch Set 15:
> > 
> > If we're already redesigning the whole command anyway, why don't we change it so that it will return records for all chips of the requested type? That would be more efficient (just one command in critical path rather than number of TCPCs + 1) and probably more useful (not all chip types we might add here in the future will necessarily have an out-of-band mechanism to check how many instances exist).
> 
> SGTM. That's more or less PS1 plus bus_type.

The caller in depthcharge that needs to know the bus information (from this lookup) is different from knowing information about PD chips and matching. Once a PD chip has been matched, then the specific driver does the lookup for the actual bus information, it is up to the driver to determine supported buses.

If we change the command to return all of the TCPC chips now, then we need a placeholder for unsupported buses so we can ensure that the index is always correct; therefore, I think it is simpler to ask for a specific index (multiple times if need be). Also the point to where you are looking up bus info for pass thru communication, you are already in a slow section of code.

And example to think about: on ITE ECs, the first 2 USBC ports have embedded TCPC chips within the EC (so they are not i2c bus) and it is possible to have a 3rd TCPC that is on the i2c bus. If we make a query to all TCPC we would need to ensure we have 3 objects to return, but the first 2 objects will be invalid/placeholders.

Patch-set: 15
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>