Update patch set 15
Patch Set 15:

> If we change the command to return all of the TCPC chips now, then we need a placeholder for unsupported buses so we can ensure that the index is always correct;

Yes, I think we should do exactly that. How is that different from requesting that index specifically and getting a zeroed out result from that?

Also the point to where you are looking up bus info for pass thru communication, you are 
already in a slow section of code.

We still need to do one command to read the amount of TCPC chips and then N commands to read chip_info for each. Why not just have this command return I2C bus info and chip info (VID/PID) for each TCPC in one go?

> And example to think about: on ITE ECs, the first 2 USBC ports have embedded TCPC chips within the EC (so they are not i2c bus) and it is possible to have a 3rd TCPC that is on the i2c bus. If we make a query to all TCPC we would need to ensure we have 3 objects to return, but the first 2 objects will be invalid/placeholders.

This is the same with the other concept, right? We'd request chip_info for all three, and would get a zeroed out result (or an error?) for the first two.

Patch-set: 15