Update patch set 1
Patch Set 1:

(1 comment)

we should think about pros and cons of this patch.

Every platform needs to be able to support 0.0.22

Those X86 platforms where there is the interrupt pulse width problem can be configured to allow short pulses. The incremental power consumption is negligible, since it kicks in only if the AP is up, do I get it right?

OTOH we have extension of the time Cr50 spends inside ISR, which is never a good thing.

So, I suggest we do nothing unless platforms which can not tolerate short pulses come up.

And whe we really need to extend this pulse, I think we should not be just looping in a for() loop, we can start the pulse in the processing routines at safe spots where interrupt processing has already finished.

Patch-set: 1
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false