Update patch set 5
Patch Set 5:

(9 comments)

I think we are making improvements in this CLs that are good (like the macro), but I we really should so that in another CL. The first CL on the stack should be a very boring move and that's it. I would even be okay with keeping the "bad" CC_ for the CPRINTS, and then updating everything in follow up CLs. It will make it easier to see what actually changed.

Patch-set: 5
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>