NX20P348X: Retrieve Vbus state during init
When the nx20p348x driver starts up, it has a 0 value for its cached
flags.  However, the PPC may still be set up to source Vbus to a port
partner and should indicate this to any callers.

On init, set the cached Vbus source flag to -1.  If the flag is
retrieved before it's been initialized, read the PPC's switch status
register in order to determine if we are sourcing Vbus (as may happen on
sysjump with a sinking partner).

BUG=b:132110509
BRANCH=octopus
TEST=ran sysjumps with display port and hoho plugged in, ensured Vbus
was detected as needing discharged when pd_power_supply_reset ran

Change-Id: I1b4a938d628d100c602b14f01b5b2e2ce1ce8c5a
Signed-off-by: Diana Z <dzigterman@chromium.org>