stm32: Refactor regs to separate stm32h7 regs
The number of stm32 chip families and variants has outgrown a single
registers header file. The nested #ifdefs are confusing and
error prone. It is difficult to determine what macros work for a given
family and which ones just haven't been set.

This change simply moves the original registers.h content to
registers-all.h, introduces a family specific header registers-stm32h7.h,
and uses registers.h to conditionally include one or the other.
to select the original header or

BUG=none
BRANCH=none
TEST=buildall

Change-Id: Ic34588a14dd5f6cbff43e926c84794c5cca4e097
Signed-off-by: Craig Hesling <hesling@chromium.org>