stm32: Add STM32H7A3/B3 support
This adds support for the STM32H7A3 and STM32H7B3.
The chip variant name is "stm32h7a3-b3" in make files.
This also adds partial support for the STM32H7B0.

The primary differences between this and the STM32H743 is
the flash layout and clock domains/max-speed.

BRANCH=icetower
BUG=b:130296790
TEST=Tested with http://crrev.com/c/2037105

Change-Id: I0cd0ac0a428f0f9632175de277a46b58734b4274
Signed-off-by: Craig Hesling <hesling@chromium.org>