Update patch set 4
Patch Set 4: Code-Review+2

I chatted with Yuval offline. My first instinct on this API would be to have the caller give a size instead an offset. But Yuval makes a good argument that this is a less invasive change and the motion_sense is the only client of this extra "staging" pointer for now, so this keeps things simpler until/unless we have a more compelling need to redesign the queue API.

I am happy with the change given the constraints.

Patch-set: 4
Reviewer: Gerrit User 1287098 <1287098@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2